(TBD) Autorequire based on resource title not ref
Previously, the acl module was returning a list of autorequires that included
the type, e.g. User[Administrator]. This results in a call to
Puppet::Resource(:user, 'User[Administrator']), which due to PUP-3177 is no
longer allowed.

It is actually not necessary to specify the type in the list of autorequired
resources, because that is controlled by the `autorequire(:user)` call.

This commit changes the autorequires blocks to return resource titles, omitting
the type.